backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Async HTTP
httpx==0.25.2
aiohttp==3.9.1

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0

# Data validation
pydantic==2.5.2

# Google Earth Engine
earthengine-api==0.1.384

# Fuzzy matching
rapidfuzz==3.5.2

# Environment and config
python-dotenv==1.0.0

# Reddit API (optional)
praw==7.7.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Type checking
mypy==1.7.1